+++ react/docs.md
@@ -1,12 +1,13 @@
 # KeysignPay
 
-A react component for setting up a payment page for tnbc purchases.
+> A react component written in typescript that makes it easy to set up a payment page for tnbc purchases.
 
-## Example
+![Keysign Payment Page Screenshot](./keysign_pay_img.png)
+
+## Usage Example
 
 ```ts
 import React, { useState } from 'react';
-import './App.css';
 import KeysignPay from 'keysign-pay-react';
 
 function App() {
@@ -40,24 +41,24 @@ function App() {
 
 ### KeysignPay
 
-| Property      | Description                                                                             | Type                                              | Default                  |
-| ------------- | --------------------------------------------------------------------------------------- | ------------------------------------------------- | ------------------------ |
-| accountNumber | The account number receiving the tnbc when someone purchases                            | string                                            |                          |
-| amount        | The amount of the item in tnbc                                                          | number                                            |                          |
-| node          | The ip address or domain of the node you want to process the transaction                | string                                            | https://bank.keysign.app |
-| onFailed      | Specify a function that will be called when the transaction fails                       | function(result: [KeysignResult](#keysignresult)) |                          |
-| onPageClose   | Specify a function that will determine what happens when a user closes the payment page | function()                                        |                          |
-| onSuccess     | Specify a function that will be called when the transaction is complete                 | function(result: [KeysignResult](#keysignresult)) |                          |
-| title         | The name of the product, subscription or service you are offering                       | string                                            |                          |
-| visible       | Displays the payment page                                                               | boolean                                           | false                    |
+| Property        | Description                                                                                                                  | Type       | Default                  |
+| --------------- | ---------------------------------------------------------------------------------------------------------------------------- | ---------- | ------------------------ |
+| `accountNumber` | The account number receiving the tnbc when someone purchases                                                                 | `string`   |                          |
+| `amount`        | The amount of the item in tnbc                                                                                               | `number`   |                          |
+| `node`          | The ip address or domain of the node you want to process the transaction                                                     | `string`   | https://bank.keysign.app |
+| `onFailed`      | Specify a function that will be called when the transaction fails. **(result: [KeysignResult](#keysignresult)) => void**     | `Function` |                          |
+| `onPageClose`   | Specify a function that will determine what happens when a user closes the payment page. **()=>void**                        | `Function` |                          |
+| `onSuccess`     | Specify a function that will be called when the transaction is complete. **(result: [KeysignResult](#keysignresult))=>void** | `Function` |                          |
+| `title`         | The name of the product, subscription or service you are offering                                                            | `string`   |                          |
+| `visible`       | Displays the payment page                                                                                                    | `boolean`  | `true`                   |
 
 ### KeysignResult
 
-| Property   | Description                                                                                     | Type          |
-| ---------- | ----------------------------------------------------------------------------------------------- | ------------- |
-| data       | The request data made to keysign's payment portal                                               | RequestData   |
-| error      | Type of error                                                                                   | string        |
-| message    | Keysign's portal message                                                                        | string        |
-| result     | The [Block](https://developer.thenewboston.com/api/bank-api/blocks) from a complete transaction | Block \| null |  |
-| request_id | Id of the request                                                                               | number        |
-| success    | Returns true if the transactions was complete and false if it failed                            | boolean       |
+| Property     | Description                                                                                     | Type              |
+| ------------ | ----------------------------------------------------------------------------------------------- | ----------------- |
+| `data`       | The request data made to keysign's payment portal                                               | `RequestData`     |
+| `error`      | Type of error                                                                                   | `string`          |
+| `message`    | Keysign's portal message                                                                        | `string`          |
+| `result`     | The [Block](https://developer.thenewboston.com/api/bank-api/blocks) from a complete transaction | `Block` \| `null` |  |
+| `request_id` | Id of the request                                                                               | `number`          |
+| `success`    | Returns true if the transactions was complete and false if it failed                            | `boolean`         |
