+++ react/src/style.css
@@ -16,9 +16,6 @@
 }
 
 @layer components {
-  .mask {
-    @apply bg-opacity-5 bg-black;
-  }
   .default-button {
     @apply text-black h-12 py-2 px-4 rounded-2xl flex  justify-center 
     items-center gap-3 font-semibold w-full;
